Added functionality for pins on MapView

+++ OnTheMap/UdacityStudentLocation.swift
@@ -16,8 +16,8 @@ struct StudentLocation {
     let lastName: String?
     let mapString: String?
     let mediaURL: String?
-    let latitude: Float?
-    let longitude: Float?
+    let latitude: Double?
+    let longitude: Double?
     let createdAt: String?
     let updatedAt: String?
     
@@ -29,8 +29,8 @@ struct StudentLocation {
         lastName = dictionary["lastName"] as? String
         mapString = dictionary["mapString"] as? String
         mediaURL = dictionary["mediaURL"] as? String
-        latitude = dictionary["latitude"] as?  Float
-        longitude = dictionary["longitude"] as? Float
+        latitude = dictionary["latitude"] as? Double
+        longitude = dictionary["longitude"] as? Double
         createdAt = dictionary["createdAt"] as? String
         updatedAt = dictionary["updatedAt"] as? String
